Reduction in use of gcc/g++ in SEL
This commit provides for the following:

1. Removal of the -m32 flags from unit tests.  These previously
necessitated the installation of the gcc-multilib package.

2. Removal of references to clang-3.7.

3. Removal of explicit installation of gcc/g++.  However, these
are still installed as prerequisites for both the build-essentials
and clang packages.

4. Removal of gcc/g++ from most areas of SEL.  gcc is stil utilized
as the default ASM linker (*asm-linker*) as replacement with clang
here led to several unit test failures related to clang being unable
to parse some constructs in the hard-coded test/etc/asm-test/grep.s.*
files.  I'm a bit less familiar with this sections of the code, so
I was hesistant to change things here.  Additionally, existing
compilation fixer tests specific to gcc have been preserved
for now.