Fix test-ccl job
In the "before_script", copy the current working directory into
the quicklisp/local-projects directory instead of creating a symlink.
This allows ccl to execute successfully, and does not appear to
have an impact on performance (the job using copy instead of
creating a symlink was slightly faster by wall time, but this
is likely due to stochastic differences between unit tests runs).