Add nightly stage to build Docker image
This should provide some balance between having an up-to-date Docker
image and not requiring a full rebuild with each push.  If a client
needs to rebuild the Docker image, they may do so manually or by
using the repository's schedules page which has a "play" button
to force execution of the scheduler.