Initial implementation of python software objects
Implements the core functionality of the software-evolution-library
for files written in python.  The core functionality is supported
by the command line script `python-astdump`, a thin wrapper
around the python AST library [1] which dumps the ASTs as JSON.
The `python-astdump` script requires python 3.8 or higher; please note
this is not the default installed by the package manager for
ubuntu-18.04 and lower.

Beyond this, the yapf [2] command line python code formatter is
required to support code formatting.  This tool may be installed
with pip.

As with JavaScript software objects, python software objects are a
parsed AST representation.  All of the generic methods for parseable
software objects have been implemented, allowing for a rich variety
of manipulations and tests.

Every AST in a python software object contains an AST class field with
the AST type (e.g. FunctionDef, Assign).  Additional fields in the python
AST representation can be found in an association list on the AST's
annotations field.  Unfortunately, these additional fields do not appear
in the official documentation, although an external AST reference
contains some details [3].

Beyond the unit tests, which instrument every AST type listed in the
python documentation, the python representation was tested to ensure
parsing completed successfully against 25 popular python projects [4].
Parsing was successful for all but two projects, both of which were
due to invalid syntax in these projects.

[1] https://docs.python.org/3/library/ast.html
[2] https://github.com/google/yapf
[3] https://greentreesnakes.readthedocs.io/en/latest/
[4] https://hackernoon.com/50-popular-python-open-source-projects-on-github-in-2018-c750f9bf56a0