Fixing almost all python tests
This changeset updates python to follow the same general framework
as JavaScript for parsing, using `convert` functions to take the
output from `python-astdump` and translating it into ASTs.  As with
JavaScript, there are fixups required on the output reported from
`python-astdump` to allow for the creation of ASTs which respect
certain SEL invariants (e.g. idempotent reads/writes of source text).
These fixups are documented and fairly straightforward.