Always use child-slots directly from acorn/python
When the concatenated child slots are not guaranteed to return
children in textual order, add a :child-order annotation to the
AST defining the textual sorted order of the children.  This
commit adds a `sorted-children` method which utilizes this
information to return the children in textual sorted order,
a requirement for printing the AST's source text.  Previously,
we concatenated and sorted the child slots into a single field
in this case; this commit preserves the original child slots.