Adding default value for interleaved text
By default, if not specified, we will populate the interleaved text
field with a list of empty strings.  The list will contain N + 1
elements, where N is the number of non-nil AST children.

This commit also adds an assertion to `source-text` ensuring the AST
has the proper amount of interleaved text.