Add copy wrapper to fixup interleaved-text
Add an around method to copy for JavaScript or python ASTs to
fixup the interleaved text field and child order annotation
after insertion and deletion.  The wrapper ensures the behavior
does not changes versus the prior implementation with a flat
list of AST children interleaved with text; however, it does not
address any issues with the prior approach.