Fix tree-sitter handling of child slot order
Switched from using a hash table to using a set for determining
if a rule had already been visited. This addresses a potential
issue with visiting a rule more than once, but not in a cyclical
manner.

Fixed which `SYMBOL' rules are visited. Previously, every `SYMBOL'
would be followed. Instead, every `SYMBOL' starting with an underscore,
'_', should be followed as its treated specially with rules and nodes.