Initial cut of sel python library
This commit finishes the existing implementation and hooks together
the `tree-sitter-interface` binary with the python library.
Additionally, this commit finishes defining unit tests for the
functionality currently exposed, including modifications
to the `python-check` Makefile target to allow for successful
testing using `make python-check`.  Currently, there are 16
unit tests of the interface which pass for both CCL and SBCL
builds of the `tree-sitter-interface`.

For now, some functions in tree-sitter-interface.lisp which
did not have unit tests nor methods defined in sel.py and
did not appear to work properly upon test were dropped.
These can be resurrected later, if required.