Add ability to work around indentation edge case
There wasn't an obvious generalize-able solution to getting
python-block working while keeping everything else working. This was
due to python-block being a purely virtual construct whereas blocks in
other languages generally have a curly brace attached to them. This also
has the benefit of keeping the indentation on the correct AST for Python.