Improve message when tree-sitter-interface crashes
An error is now always triggered; previously, an error would only
be trigger if assertions were enabled.  Additionally, the message
now refers to the "tree-sitter-interface" command in all cases;
previously, if the process was spawn from the python package,
the message would refer to an oblique path to the command.
Finally, if stderr if populated, print the stderr; otherwise,
elide any mention of stderr and only note the interface crashed.