Sockets for tree-sitter-interface communication
This will allow for the use of ProcessPool and ProcessPoolExecutor
objects with the tree-sitter-interface.  For true concurrent
processing in python, these process pool objects must be utilized.
However, because process pools spawn new python subprocesses, the
tree-sitter-interface object cannot be shared with them, and thus,
we cannot write to stdin and stdout of the tree-sitter-interface
object in these python subprocesses.