Fix deadlock in garbage collection
A deadlock can occur when garbage collection is initiated while
in a section guarded by a multiprocessing lock.  To avoid this, we
place the ASTs to be garbage collected in a queue that is flushed
to the LISP subprocess periodically when the garbage collection
queue size reaches a certain threshold.  This has a side benefit
of being more efficient than sending ASTs to be deleted individually
to the LISP subprocess.