Improvements to `pick-2-moveable`
1. Use `pick-2-swappable` when picking the AST to move.
This ensures the AST being moved is not a subtree of the
old AST being replaced.  Additionally, it avoids various
issues with `pick-2-replaceable` and the `tree-copy` method.

2. After finding the ASTs involved in the move, ensure we
can cut the second AST in the picked selection.  The second
AST is location of the AST being moved and will be cut in
the result.