Two fixes to infer-type in cases that were causing errors.
In the first, x.y when the type of x is not defined
was causing an error.  Make it return nil if the field-table
cannot be found.

In the second, the & operator on pointer variables was causing
creation of an unprintable type descriptor.  In this case,
return a type descriptor with the wrong number of **.  The code
was wrong anyway; see TODOs.