Add cl-yaml and cl-yaml-project objects
The genome of these is the data structures produced
by cl-yaml:parse.  Round tripping is not implemented,
and would not be byte equivalent if it were.  For byte
equivalence, a yaml tree-sitter parser should be used.
Those software objects will go by the name yaml when
they are added.

Also, cut back the CFLAGS when compiling tree-sitter
to allow it to be compiled on older clangs without
failing due to -Werror.